# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountScoreBand": {
        "description": "`account_score.band`.",
        "enum": [
          "HOT",
          "WARM",
          "COLD"
        ],
        "title": "AccountScoreBand",
        "type": "string"
      },
      "AccountScoreStanding": {
        "description": "`account_score.standing`.",
        "enum": [
          "RANKED",
          "BELOW_FIT",
          "DISQUALIFIED",
          "CUSTOMER"
        ],
        "title": "AccountScoreStanding",
        "type": "string"
      },
...
      "DisqualifierOverrideStatus": {
        "description": "`disqualifier_override.status`.",
        "enum": [
          "ACTIVE",
          "REVOKED"
        ],
        "title": "DisqualifierOverrideStatus",
        "type": "string"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The `{\"error\": {...}}` shape of [Conventions](/architecture/interfaces.md#conventions),\nnamed `ErrorEnvelope` in the OpenAPI document.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "EvidenceView": {
        "additionalProperties": false,
        "description": "[`EvidenceView`](/architecture/interfaces.md#evidenceview).",
        "properties": {
          "document": {
            "$ref": "#/components/schemas/FindingDocument"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "purged": {
            "title": "Purged",
            "type": "boolean"
          },
          "quote_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote End"
          },
          "quote_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Start"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          }
        },
        "required": [
          "finding_id",
          "document",
          "section",
          "purged",
          "excerpt",
          "quote_start",
          "quote_end"
        ],
        "title": "EvidenceView",
        "type": "object"
      },
...
      "FindingDecidedBy": {
        "description": "`finding.decided_by`.",
        "enum": [
          "CLASSIFIER",
          "LLM"
        ],
        "title": "FindingDecidedBy",
        "type": "string"
      },
...
      "FindingStatus": {
        "description": "`finding.status`.",
        "enum": [
          "ACTIVE",
          "SUPERSEDED",
          "REJECTED"
        ],
        "title": "FindingStatus",
        "type": "string"
      },
      "FindingStrength": {
        "description": "`finding.strength`; reused by `classification.strength` and\n`evaluation_item.expected_strength`.",
        "enum": [
          "NONE",
          "WEAK",
          "MEDIUM",
          "STRONG"
        ],
        "title": "FindingStrength",
        "type": "string"
      },
      "FindingView": {
        "additionalProperties": false,
        "description": "[`FindingView`](/architecture/interfaces.md#findingview), the response of `API-47`. Also\nplan task 13's `API-42` response shape; that task imports this model and must not define a\nsecond one.",
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "decided_by": {
            "$ref": "#/components/schemas/FindingDecidedBy"
          },
          "document": {
            "$ref": "#/components/schemas/FindingViewDocument"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FindingViewFeedback"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "option": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FindingViewOption"
              },
              {
                "type": "null"
              }
            ]
          },
          "points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "question": {
            "$ref": "#/components/schemas/FindingViewQuestion"
          },
          "question_revision": {
            "title": "Question Revision",
            "type": "integer"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "quote_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote En"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "service_id": {
            "format": "uuid",
            "title": "Service Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "strength": {
            "$ref": "#/components/schemas/FindingStrength"
          }
        },
        "required": [
          "id",
          "account_id",
          "service_id",
          "question",
          "question_revision",
          "confidence",
          "quote",
          "quote_en",
          "rationale",
          "observed_at",
          "strength",
          "decided_by",
          "status",
          "option",
          "document",
          "points",
          "feedback"
        ],
        "title": "FindingView",
        "type": "object"
      },
      "FindingViewDocument": {
        "additionalProperties": false,
        "description": "`FindingView.document`.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "plugin_code": {
            "$ref": "#/components/schemas/SourcePluginCode"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "source_type": {
            "$ref": "#/components/schemas/DocumentSourceType"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "url",
          "source_type",
          "plugin_code",
          "language",
          "published_at"
        ],
        "title": "FindingViewDocument",
        "type": "object"
      },
      "FindingViewFeedback": {
        "additionalProperties": false,
        "description": "`FindingView.feedback`: the in-force [`finding_feedback`]\n(/architecture/sql-store.md#finding_feedback).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/FindingFeedbackVerdict"
          }
        },
        "required": [
          "verdict",
          "user_name",
          "created_at"
        ],
        "title": "FindingViewFeedback",
        "type": "object"
      },
      "FindingViewOption": {
        "additionalProperties": false,
        "description": "`FindingView.option`; `CHOICE` questions only.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label"
        ],
        "title": "FindingViewOption",
        "type": "object"
      },
...
      "Override": {
        "additionalProperties": false,
        "description": "[`Override`](/architecture/interfaces.md#override).",
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "created_by_name": {
            "title": "Created By Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "revoked_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked By Name"
          },
          "rule_key": {
            "title": "Rule Key",
            "type": "string"
          },
          "rule_label": {
            "title": "Rule Label",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "service_id": {
            "title": "Service Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DisqualifierOverrideStatus"
          }
        },
        "required": [
          "id",
          "account_id",
          "service_id",
          "rule_key",
          "note",
          "rule_label",
          "status",
          "created_by_name",
          "created_at",
          "revoked_by_name",
          "revoked_at",
          "run_id"
        ],
        "title": "Override",
        "type": "object"
      },
      "OverrideCreate": {
        "additionalProperties": false,
        "description": "[`OverrideCreate`](/architecture/interfaces.md#overridecreate).",
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "rule_key": {
            "title": "Rule Key",
            "type": "string"
          }
        },
        "required": [
          "rule_key",
          "note"
        ],
        "title": "OverrideCreate",
        "type": "object"
      },
...
      "ProspectPage": {
        "additionalProperties": false,
        "description": "[`ProspectPage`](/architecture/interfaces.md#prospectpage): `Page<ProspectRow>` plus\n`band_counts`.",
        "properties": {
          "band_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/AccountScoreBand"
            },
            "title": "Band Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProspectRow"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "band_counts"
        ],
        "title": "ProspectPage",
        "type": "object"
      },
...
      "ScoreView": {
        "additionalProperties": false,
        "description": "[`ScoreView`](/architecture/interfaces.md#scoreview).",
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "as_of": {
            "title": "As Of",
            "type": "string"
          },
          "band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountScoreBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "breakdown": {
            "$ref": "#/components/schemas/ScoreViewBreakdown"
          },
          "fit": {
            "title": "Fit",
            "type": "integer"
          },
          "intent": {
            "title": "Intent",
            "type": "integer"
          },
          "last_crm_sync": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrmSyncView"
              },
              {
                "type": "null"
              }
            ]
          },
          "lead_feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadFeedback"
              },
              {
                "type": "null"
              }
            ]
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/Override"
            },
            "title": "Overrides",
            "type": "array"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "score_id": {
            "title": "Score Id",
            "type": "string"
          },
          "scoring_version": {
            "title": "Scoring Version",
            "type": "integer"
          },
          "service_id": {
            "title": "Service Id",
            "type": "string"
          },
          "standing": {
            "$ref": "#/components/schemas/AccountScoreStanding"
          }
        },
        "required": [
          "score_id",
          "account_id",
          "service_id",
          "scoring_version",
          "as_of",
          "fit",
          "intent",
          "priority",
          "standing",
          "band",
          "rank",
          "breakdown",
          "overrides",
          "lead_feedback",
          "last_crm_sync"
        ],
        "title": "ScoreView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LeadRadar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/accounts/{id}/findings": {
      "get": {
        "description": "`API-42`.",
        "operationId": "list_findings_api_v1_accounts__id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "service_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "in": "query",
            "name": "question_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FindingView"
                  },
                  "title": "Response List Findings Api V1 Accounts  Id  Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List Findings",
        "tags": [
          "prospects-and-evidence"
        ]
      }
    },
...
    "/api/v1/accounts/{id}/scores/{service_id}": {
      "get": {
        "description": "`API-40`.",
        "operationId": "get_score_api_v1_accounts__id__scores__service_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get Score",
        "tags": [
          "prospects-and-evidence"
        ]
      }
    },
...
    "/api/v1/accounts/{id}/scores/{service_id}/overrides": {
      "post": {
        "description": "`API-44`.",
        "operationId": "create_override_api_v1_accounts__id__scores__service_id__overrides_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverrideCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Override"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create Override",
        "tags": [
          "prospects-and-evidence"
        ]
      }
    },
...
    "/api/v1/findings/{id}/evidence": {
      "get": {
        "description": "`API-43`.",
        "operationId": "get_evidence_api_v1_findings__id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "prospects-and-evidence"
        ]
      }
    },
...
    "/api/v1/overrides/{id}/revoke": {
      "post": {
        "description": "`API-45`.",
        "operationId": "revoke_override_api_v1_overrides__id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Override"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Revoke Override",
        "tags": [
          "prospects-and-evidence"
        ]
      }
    },
...
    "/api/v1/services/{id}/prospects": {
      "get": {
        "description": "`API-39`.",
        "operationId": "list_prospects_api_v1_services__id__prospects_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "standing",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AccountScoreStanding",
              "default": "RANKED"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/AccountScoreBand"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Band"
            }
          },
          {
            "in": "query",
            "name": "country_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country Code"
            }
          },
          {
            "in": "query",
            "name": "industry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "priority",
                    "intent",
                    "fit",
                    "name",
                    "last_refreshed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProspectPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List Prospects",
        "tags": [
          "prospects-and-evidence"
        ]
      }
    },
...
  }
}